feature: handle multiple connections
To achieve this, the architecture had to change. Connections were kept open, and global variables were reused, making parallell execution impossible. Thus, to fix this, the client now must Dial the server each time to send a message, closing its connection once the response has been received. As for the Server Request Handler, in order to handle simultaneous connections, a 'lock' channel has been added, adding synchronization to this component.

+++ exercise_06/mymiddleware/service/namingProxy.go
@@ -52,10 +52,10 @@ func (n NamingServiceProxy) Lookup(proxyName string) Proxy {
 	if (mappedProxy["TypeName"] == "queue") {
 		return QueueProxy{
 			HostIp: mappedProxy["HostIp"].(string),
-			HostPort: int(mappedProxy["HostPort"].(int64)),
-			RemoteObjectId:int(mappedProxy["RemoteObjectId"].(int64)),
+			HostPort: int(mappedProxy["HostPort"].(float64)),
+			RemoteObjectId:int(mappedProxy["RemoteObjectId"].(float64)),
 			TypeName: mappedProxy["TypeName"].(string),
-			QueueNumber: int(mappedProxy["QueueNumber"].(int64))}
+			QueueNumber: int(mappedProxy["QueueNumber"].(float64))}
 	} else {
 		return StackProxy{
 			HostIp: mappedProxy["HostIp"].(string),
